+++ pytsite/browser/res/assets/jquery-ui/jquery-ui-loader.js
@@ -1,4 +1,4 @@
-define(['assetman', 'jquery-ui-main'], function(assetman, jqueryUI) {
+define(['assetman', 'jquery-ui-vendor'], function(assetman, jqueryUI) {
     assetman.loadCSS('pytsite.browser@jquery-ui/jquery-ui.css');
 
     return jqueryUI;