+++ pytsite/browser/res/assets/twitter-bootstrap-tokenfield/bootstrap-tokenfield-loader.js
@@ -1,4 +1,4 @@
-define(['jquery-ui', 'twitter-bootstrap-tokenfield-main', 'assetman'], function(jqueryUI, tokenfield, assetman) {
+define(['jquery-ui', 'twitter-bootstrap-tokenfield-vendor', 'assetman'], function(jqueryUI, tokenfield, assetman) {
     assetman.loadCSS('pytsite.browser@twitter-bootstrap-tokenfield/css/bootstrap-tokenfield.css');
 
     return tokenfield;