+++ pytsite/browser/res/assets/twitter-bootstrap/twitter-bootstrap-loader.js
@@ -1,6 +1,5 @@
-define(['assetman'], function(assetman) {
+define(['assetman', 'twitter-bootstrap-vendor'], function(assetman) {
     assetman.loadCSS('pytsite.browser@twitter-bootstrap/css/bootstrap.css');
     assetman.loadCSS('pytsite.browser@twitter-bootstrap/css/add-columns.css');
     assetman.loadCSS('pytsite.browser@twitter-bootstrap/css/add-ons.css');
-    assetman.loadJS('pytsite.browser@twitter-bootstrap/js/bootstrap.js');
 });