+++ pytsite/browser/res/assets/typeahead/typeahead-loader.js
@@ -1,3 +1,3 @@
-define(['assetman', 'typeahead-main'], function(assetman) {
+define(['assetman', 'typeahead-vendor'], function(assetman) {
     assetman.loadCSS('pytsite.browser@typeahead/typeahead.css');
 });