+++ pytsite/form/__init__.py
@@ -17,9 +17,9 @@ def _init():
     tpl.register_package(__name__)
 
     assetman.register_package(__name__)
-    assetman.t_less(__name__ + '@**/*.less')
-    assetman.t_js(__name__ + '@**/*.js')
-    assetman.js_module('pytsite-form', __name__ + '@js/pytsite-form')
+    assetman.t_less(__name__ + '@**')
+    assetman.t_js(__name__ + '@**')
+    assetman.js_module('pytsite-form-module', __name__ + '@js/pytsite-form-module')
 
     router.handle(_controllers.Submit(), '/form/submit/<uid>', 'pytsite.form@submit', methods='POST')